Move withFonts logic directly into `_app`
