+++ pages/_app.js
@@ -1,19 +1,31 @@
 import App, { Container } from 'next/app';
 import Router from 'next/router';
-import ScrollUpButton from 'react-scroll-up-button';
-import ReactGA from 'react-ga';
+import FontFaceObserver from 'fontfaceobserver';
 import LogRocket from 'logrocket';
+import ReactGA from 'react-ga';
+import ScrollUpButton from 'react-scroll-up-button';
 import setupLogRocketReact from 'logrocket-react';
 import * as Sentry from '@sentry/browser';
 import Nav from 'components/Nav/Nav';
 import Footer from 'components/Footer/Footer';
 import Modal from 'components/Modal/Modal';
-import withFonts from 'decorators/withFonts/withFonts';
 import 'common/styles/globalStyles.css';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
-// eslint-disable-next-line react/prefer-stateless-function
+const fonts = [
+  {
+    fontFamily: 'Encode Sans',
+    url: 'https://fonts.googleapis.com/css?family=Encode+Sans:400,700',
+  },
+  {
+    fontFamily: 'DIN Condensed Bold',
+    // loading of this font is being handled by the @font-face rule on
+    // the global style sheet.
+    url: null,
+  },
+];
+
 class Layout extends React.Component {
   render() {
     // eslint-disable-next-line react/prop-types
@@ -32,6 +44,22 @@ class Layout extends React.Component {
 
 class OperationCodeApp extends App {
   componentDidMount() {
+    const observers = fonts.map(font => {
+      if (font.url) {
+        const link = document.createElement('link');
+        link.href = font.url;
+        link.rel = 'stylesheet'; // eslint-disable-line unicorn/prevent-abbreviations
+        document.head.append(link);
+      }
+
+      const observer = new FontFaceObserver(font.fontFamily);
+      return observer.load();
+    });
+
+    Promise.all(observers).then(() => {
+      document.documentElement.classList.add('fonts-loaded');
+    });
+
     if (isProduction) {
       Sentry.init({ dsn: process.env.SENTRY_DSN });
       LogRocket.init(`${process.env.LOGROCKET_KEY}/operation-code`);
@@ -98,4 +126,4 @@ if (!isProduction) {
   });
 }
 
-export default withFonts(OperationCodeApp);
+export default OperationCodeApp;
